pyproject.toml
[project]
name = "email-ai-assistant"
version = "0.1.0"
description = "MVP toolkit for an email-focused AI assistant with n8n workflows"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.2",
    "tabulate>=0.9.0",
    "typer>=0.9.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0"
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]
